+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Removed
 - :warning: make sure to comment out any removed options in your config
 - ``sling-refresh-peers-interval`` lowering it from the default 1 doesn't help much and if you must set it higher your node is probably too slow anyways
+- ``sling-refresh-gossmap-interval`` gossip reader is so fast now there is no good reason to keep it
 
 ### Added
 - `sling-once`: new command to rebalance a specific amount once